requirements.txt
action-tutorials-py==0.9.4
ament-copyright==0.9.8
ament-cppcheck==0.9.8
ament-cpplint==0.9.8
ament-flake8==0.9.8
ament-index-python==1.1.0
ament-lint==0.9.8
ament-lint-cmake==0.9.8
ament-package==0.9.5
ament-pep257==0.9.8
ament-uncrustify==0.9.8
ament-xmllint==0.9.8
cloudpickle==3.1.1
contourpy==1.1.1
cycler==0.12.1
demo-nodes-py==0.9.4
domain-coordinator==0.9.2
empy==4.2
examples-rclpy-executors==0.9.4
examples-rclpy-minimal-action-client==0.9.4
examples-rclpy-minimal-action-server==0.9.4
examples-rclpy-minimal-client==0.9.4
examples-rclpy-minimal-publisher==0.9.4
examples-rclpy-minimal-service==0.9.4
examples-rclpy-minimal-subscriber==0.9.4
Farama-Notifications==0.0.4
filelock==3.16.1
fonttools==4.56.0
fsspec==2025.3.0
gymnasium==1.1.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
jinja2==3.1.6
kiwisolver==1.4.7
lark-parser==0.12.0
launch==0.10.10
launch-ros==0.11.7
launch-testing==0.10.10
launch-testing-ros==0.11.7
launch-xml==0.10.10
launch-yaml==0.10.10
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
opencv-python==4.11.0.86
osrf-pycommon==0.1.11
packaging==24.2
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
quality-of-service-demo-py==0.9.4
ros2action==0.9.13
ros2bag==0.3.11
ros2cli==0.9.13
ros2component==0.9.13
ros2controlcli==0.11.0
ros2doctor==0.9.13
ros2interface==0.9.13
ros2launch==0.11.7
ros2lifecycle==0.9.13
ros2multicast==0.9.13
ros2node==0.9.13
ros2param==0.9.13
ros2pkg==0.9.13
ros2run==0.9.13
ros2service==0.9.13
ros2topic==0.9.13
rosidl-runtime-py==0.9.1
rpyutils==0.2.0
rqt-action==1.0.2
rqt-console==1.1.2
rqt-graph==1.1.3
rqt-gui==1.1.2
rqt-gui-py==1.1.2
rqt-msg==1.1.1
rqt-plot==1.1.1
rqt-publisher==1.3.0
rqt-py-console==1.0.2
rqt-reconfigure==1.0.8
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-top==1.0.2
rqt-topic==1.3.0
six==1.17.0
sros2==0.9.5
sympy==1.13.3
teleop-twist-keyboard==2.3.2
topic-monitor==0.9.4
torch==2.4.1
triton==3.0.0
turtlebot3-example==2.1.1
turtlebot3-teleop==2.1.1
typing-extensions==4.12.2
zipp==3.20.2
